Added container support for the lua-nginx-module test suite:
The image is created by running docker build in the root directory of the
project. It contains all the dependencies of the environment. This includes not
only nginx, but all the modules that are needed to be compiled with it,
and the services that are required to be run in the testing environment.

In addition it contains the script container_build.sh, which maps to
/nginx/nginx/test_build.sh in the image, that configures, builds and installs
Nginx. It also contains the script run_test_dependencies.sh which runs the
required services for some of the tests.

In order to build the image, run:
  docker build -t IMAGE_NAME .

In order to start running tests, you can run from the project's root directory:
  docker run -it --rm=true -v $(pwd):/nginx/modules/lua-nginx-module/ nginx-lua-tests /bin/bash
Then, inside the container, /nginx/nginx/ is Nginx's directory, you can build it
using the test_build.sh script. Run the tests by first running the script
run_test_dependencies.sh (it's in PATH) and then by cd'ing the directory
/nginx/modules/lua-nginx/module and by running prove.